Added API Instance

+++ app/src/main/java/com/main/data/API/dto/RecipeClass.kt
@@ -11,5 +11,5 @@ data class RecipeClass(
     @SerializedName("ingredients") @Expose
     val ingredients: List<IngredientClass>,
     @SerializedName("totalWeigh") @Expose
-    val totalWeight: Float
+    val totalWeight: Double
 )
